shifted auto scoring page from components to correct folder, pages
